refactor into a helper
